+++ BriansUsbQuizBoxApi/Helpers/DebugHelpers.cs
@@ -9,13 +9,24 @@ namespace BriansUsbQuizBoxApi.Helpers
     static public class DebugHelpers
     {
         /// <summary>
-        /// Formatted debug write line with timestamp
+        /// Debug write line with timestamp
         /// </summary>
         /// <param name="message">Message to write</param>
         [Conditional("DEBUG")]
         static public void WriteLine(string message)
         {
             Debug.WriteLine(DateTime.Now.ToString("HH:mm:ss.fff") + " --- " + message);
+        }
+
+        /// <summary>
+        /// Formatted debug write line with timestamp
+        /// </summary>
+        /// <param name="format">Message to format and write</param>
+        /// <param name="args">Message format arguments</param>
+        [Conditional("DEBUG")]
+        static public void WriteLine(string format, params object[] args)
+        {
+            Debug.WriteLine(DateTime.Now.ToString("HH:mm:ss.fff") + " --- " + format, args);
         }
     }
 }