+++ CHANGELOG.md
@@ -22,3 +22,10 @@ Track changes made to the API.
 - Enhanced test application
 - Updated projects to .NET Core 9 and latest dependencies (including HidSharp)
 - Enhanced 'Read Me' sample code
+
+## Version 1.1.0-beta.2 (1.1.10.0)
+
+- Second beta of Kirkman Basic Quizbox Plus support
+- Documentation improvements for new support
+- General code cleanup
+- Update project to .NET Core 10